// dashboard/lib/types/credentials.ts
export enum HostType {
  eu_shared = 'https://api.tinybird.co',
  us_east = 'https://api.us-east.tinybird.co',
  aws_us_east = 'https://api.us-east.aws.tinybird.co',
  aws_us_west_2 = 'https://api.us-west-2.aws.tinybird.co',
  aws_eu_central_1 = 'https://api.eu-central-1.aws.tinybird.co',
  aws_eu_west_1 = 'https://api.eu-west-1.aws.tinybird.co',
  gcp_europe_west2 = 'https://api.europe-west2.gcp.tinybird.co',
  gcp_europe_west3 = 'https://api.europe-west3.gcp.tinybird.co',
  gcp_us_east4 = 'https://api.us-east4.gcp.tinybird.co',
  gcp_northamerica_northeast2 = 'https://api.northamerica-northeast2.gcp.tinybird.co',
  local = 'http://localhost:7181',
  dev = 'http://127.0.0.1:8001',
  other = 'other',
}